+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VBottomSheetDemoView.swift
@@ -144,7 +144,7 @@ struct VBottomSheetDemoView: View {
         VCheckBox(
             uiModel: {
                 var uiModel: VCheckBoxUIModel = .init()
-                uiModel.layout.titleLineLimit = 1
+                uiModel.layout.titleType = .singleLine
                 return uiModel
             }(),
             isOn: .init(