+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VModalDemoView.swift
@@ -87,7 +87,7 @@ struct VModalDemoView: View {
         VCheckBox(
             uiModel: {
                 var uiModel: VCheckBoxUIModel = .init()
-                uiModel.layout.titleLineLimit = 1
+                uiModel.layout.titleType = .singleLine
                 return uiModel
             }(),
             isOn: .init(