+++ Sources/VComponents/Components/Buttons/Plain Button/VPlainButton.swift
@@ -117,6 +117,7 @@ public struct VPlainButton<Label>: View where Label: View {
     
     private func labelTitleComponent(title: String) -> some View {
         VText(
+            minimumScaleFactor: uiModel.layout.titleMinimumScaleFactor,
             color: uiModel.colors.title.value(for: internalState),
             font: uiModel.fonts.title,
             text: title