+++ Sources/VComponents/Components/Buttons/Rounded Button/VRoundedButtonUIModel.swift
@@ -45,6 +45,9 @@ public struct VRoundedButtonUIModel {
         /// Label margins. Defaults to `3`s.
         public var labelMargins: LabelMargins = .init(3)
         
+        /// Title minimum scale factor. Defaults to `0.75`.
+        public var titleMinimumScaleFactor: CGFloat = 0.75
+        
         /// Icon size. Defaults to `20` by `20`.
         public var iconSize: CGSize = primaryButtonReference.layout.iconSize