+++ Sources/VComponents/Components/Buttons/Rounded Labeled Button/VRoundedLabeledButtonUIModel.swift
@@ -64,6 +64,9 @@ public struct VRoundedLabeledButtonUIModel {
         /// Title label type. Defaults to `multiline` of `1...2` lines.
         public var titleLabelType: VTextType = .multiLine(alignment: .center, lineLimit: 1...2)
         
+        /// Title label minimum scale factor. Defaults to `0.75`.
+        public var titleLabelMinimumScaleFactor: CGFloat = 0.75
+        
         // MARK: Initializers
         /// Initializes sub-model with default values.
         public init() {}